requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
motor==3.3.2
pymongo==4.6.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# PDF Processing
PyPDF2==3.0.1

# LLM Integration
google-generativeai==0.3.2

# Configuration & Environment
python-dotenv==1.0.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# Logging
loguru==0.7.2

# Development dependencies (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Production dependencies
gunicorn==21.2.0